Serve from the path of the public hostname
Passing public-hostname with a non-empty path will use that path as the
prefix for the HTTP handler.

This is technically a breaking change, in that anybody who was using a
non-empty path behind a reverse proxy must have had it configured to
remove the path prefix on that end. Now they need to remove that.
However, that's generally a brittle thing to do; it's simpler all around
for the backend server to know what its root path should be.

Moreover, Firebase Hosting doesn't have such a capability, so it would
require running an additional backend server.