Progressbar writer
Use the writer from the logger to write the progressbar output to.
The logger is available in every function that uses a progressbar but it is often un(der)used.
As the default output writer for loggers is os.Stderr as which is the default for the progressbar as well, this won't change anything for default logger users, but it enables users of the library to redirect the progressbar output.